requirements.txt
alembic==1.0.10
amqp==2.4.2
billiard==3.6.0.0
celery==4.3.0
Click==7.0
colorama==0.4.1
Flask==1.0.3
Flask-Migrate==2.5.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
gunicorn==19.9.0
honcho==1.0.1
itsdangerous==1.1.0
Jinja2==2.10.1
kombu==4.5.0
Mako==1.0.10
MarkupSafe==1.1.1
marshmallow==2.19.2
mysqlclient==1.4.2
python-dateutil==2.8.0
python-editor==1.0.4
python-envcfg==0.2.0
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.3
vine==1.3.0
Werkzeug==0.15.4
